Added log messages to parts of dojo-adapter that may not be working yet.
Turned off try-catch error handlers in jsPlumbUtil.wrap to aid debugging.
It would be good to have a jsPlumb debug flag to turn things like this on/off.

initDraggable is now attaching the drag callback to dojo/dnd/onMove.
However, the callback is showing a number of errors, possibly due to the
fact that the callback is now being given a dojo/dnd mover object as an
argument.

Still need to work on initDroppable.